19: Moving Navigation component to folder
